Enhancement performance is now applied to both cluster and thread pools
